Add files via upload
This is the first phase of the project. Here, all the audio files are stored in a data frame and converted in .csv format. Then Librosa is used for detecting the emotions of the speeches given by different actors. A spectrogram and histogram of the audio have also been plotted.